fixes #7297 - corrected name of DNS namespace in virsh error handlers
Refactored DNS and DHCP specifics back into the respective modules

(cherry picked from commit b32f2f777c5e23119f3daced9c96432d08d5bef7)
