+++ modules/dns/providers/virsh.rb
@@ -1,4 +1,5 @@
 require "proxy/virsh"
+require 'rexml/document'
 
 module Proxy::Dns
   class Virsh < Record
@@ -36,6 +37,34 @@ module Proxy::Dns
       else
         logger.warn "not deleting #{@type} record for #{@fqdn} (unsupported)"
       end
+    end
+
+    private
+
+    def find_ip_for_host host
+      doc = REXML::Document.new xml = dump_xml
+      doc.elements.each("network/dns/host/hostname") do |e|
+        if e.text == host
+          return e.parent.attributes["ip"]
+        end
+      end
+      raise Proxy::Dns::Error.new("Cannot find DNS entry for #{host}")
+    rescue Exception => e
+      msg = "DNS virsh provider error: unable to retrive virsh info: #{e}"
+      logger.error msg
+      logger.debug xml if defined?(xml)
+      raise Proxy::Dns::Error, msg
+    end
+
+    def virsh_update_dns command, hostname, ip
+      hostname = escape_for_shell(hostname)
+      ip = escape_for_shell(ip)
+      net = escape_for_shell(network)
+      virsh "net-update", net, command, "dns-host",
+        "--xml", "'<host ip=\"#{ip}\"><hostname>#{hostname}</hostname></host>'",
+        "--live", "--config"
+    rescue Proxy::Virsh::Error => e
+      raise Proxy::Dns::Error, "Failed to update DNS: #{e}"
     end
 
   end
